+++ episodic/orchestration/__init__.py
@@ -1,5 +1,7 @@
 """Structured planning and tool execution for generation orchestration."""
 
+from episodic.orchestration._checkpoint_resume import resume_generation_orchestration
+from episodic.orchestration._graph_state import GenerationGraphState
 from episodic.orchestration.checkpoints import InMemoryCheckpointStore
 from episodic.orchestration.generation import (
     ActionExecutionResult,
@@ -30,11 +32,7 @@ from episodic.orchestration.generation import (
     build_generation_result,
     build_workflow_step_idempotency_key,
 )
-from episodic.orchestration.langgraph import (
-    GenerationGraphState,
-    build_generation_orchestration_graph,
-    resume_generation_orchestration,
-)
+from episodic.orchestration.langgraph import build_generation_orchestration_graph
 
 __all__ = [
     "ActionExecutionResult",
